Refactor IsomorphicGraphMapping
