+++ jgrapht-core/src/main/java/org/jgrapht/alg/isomorphism/IsomorphicGraphMapping.java
@@ -17,7 +17,8 @@
  */
 package org.jgrapht.alg.isomorphism;
 
-import org.jgrapht.*;
+import org.jgrapht.Graph;
+import org.jgrapht.GraphMapping;
 
 import java.util.*;
 
@@ -25,16 +26,19 @@ import java.util.*;
  * This class represents a GraphMapping between two (subgraph)isomorphic graphs. In the subgraph
  * isomorphic case, the second one is assumed to be a subgraph of the first one.
  *
+ * @author Fabian Späh
+ * @author Alexandru Valeanu
+ *
  * @param <V> the type of the vertices
  * @param <E> the type of the edges
  */
-public class IsomorphicGraphMapping<V, E>
-    implements
-    GraphMapping<V, E>
-{
-    GraphOrdering<V, E> g1, g2;
+public class IsomorphicGraphMapping<V, E> implements GraphMapping<V, E> {
 
-    int[] core1, core2;
+    private final Map<V, V> forwardMapping;
+    private final Map<V, V> backwardMapping;
+
+    private final Graph<V, E> graph1;
+    private final Graph<V, E> graph2;
 
     /**
      * Construct a new isomorphic graph mapping
@@ -44,64 +48,107 @@ public class IsomorphicGraphMapping<V, E>
      * @param core1 the mapping as array (forwards)
      * @param core2 the mapping as array (backwards)
      */
-    public IsomorphicGraphMapping(
-        GraphOrdering<V, E> g1, GraphOrdering<V, E> g2, int[] core1, int[] core2)
-    {
-        this.g1 = g1;
-        this.g2 = g2;
-        this.core1 = core1.clone();
-        this.core2 = core2.clone();
+    public IsomorphicGraphMapping(GraphOrdering<V, E> g1, GraphOrdering<V, E> g2, int[] core1, int[] core2) {
+        this.graph1 = g1.getGraph();
+        this.graph2 = g2.getGraph();
+
+        this.forwardMapping = new HashMap<>(this.graph1.vertexSet().size());
+        this.backwardMapping = new HashMap<>(this.graph1.vertexSet().size());
+
+        for (V v: graph1.vertexSet()){
+            int vNumber = g1.getVertexNumber(v);
+            int uNumber = core1[vNumber];
+
+            if (uNumber != VF2State.NULL_NODE) {
+                forwardMapping.put(v, g2.getVertex(uNumber));
+            }
+        }
+
+        for (V v: graph2.vertexSet()){
+            int vNumber = g2.getVertexNumber(v);
+            int uNumber = core2[vNumber];
+
+            if (uNumber != VF2State.NULL_NODE) {
+                backwardMapping.put(v, g1.getVertex(uNumber));
+            }
+        }
+    }
+
+    /**
+     * Construct a new isomorphic graph mapping.
+     *
+     * @param forwardMapping the mapping from graph1 to graph2
+     * @param backwardMapping the mapping from graph2 to graph1 (inverse of forwardMapping)
+     * @param graph1 the first graph
+     * @param graph2 the second graph
+
+     */
+    public IsomorphicGraphMapping(Map<V, V> forwardMapping, Map<V, V> backwardMapping, Graph<V, E> graph1, Graph<V, E> graph2) {
+        this.forwardMapping = Objects.requireNonNull(forwardMapping);
+        this.backwardMapping = Objects.requireNonNull(backwardMapping);
+
+        this.graph1 = Objects.requireNonNull(graph1);
+        this.graph2 = Objects.requireNonNull(graph2);
     }
 
     @Override
-    public V getVertexCorrespondence(V v, boolean forward)
-    {
-        GraphOrdering<V, E> firstGraph, secondGraph;
-        int[] core;
-
-        if (forward) {
-            firstGraph = g1;
-            secondGraph = g2;
-            core = core1;
-        } else {
-            firstGraph = g2;
-            secondGraph = g1;
-            core = core2;
-        }
-
-        int vNumber = firstGraph.getVertexNumber(v), uNumber = core[vNumber];
-
-        if (uNumber == VF2State.NULL_NODE) {
-            return null;
-        }
-
-        return secondGraph.getVertex(uNumber);
+    public V getVertexCorrespondence(V v, boolean forward) {
+        if (forward)
+            return forwardMapping.get(v);
+        else
+            return backwardMapping.get(v);
     }
 
     @Override
-    public E getEdgeCorrespondence(E e, boolean forward)
-    {
-        GraphOrdering<V, E> firstGraph, secondGraph;
-        int[] core;
-
+    public E getEdgeCorrespondence(E e, boolean forward) {
         if (forward) {
-            firstGraph = g1;
-            secondGraph = g2;
-            core = core1;
+            V u = graph1.getEdgeSource(e);
+            V v = graph1.getEdgeTarget(e);
+
+            return graph2.getEdge(forwardMapping.get(u), forwardMapping.get(v));
+
         } else {
-            firstGraph = g2;
-            secondGraph = g1;
-            core = core2;
-        }
+            V u = graph2.getEdgeSource(e);
+            V v = graph2.getEdgeTarget(e);
 
-        int[] eNumbers = firstGraph.getEdgeNumbers(e);
-        if ((core[eNumbers[0]] == VF2State.NULL_NODE)
-            || (core[eNumbers[1]] == VF2State.NULL_NODE))
-        {
-            return null;
+            return graph1.getEdge(backwardMapping.get(u), backwardMapping.get(v));
         }
+    }
 
-        return secondGraph.getEdge(core[eNumbers[0]], core[eNumbers[1]]);
+    /**
+     * Get an unmodifiable version of the forward mapping function.
+     *
+     * @return the unmodifiable forward mapping function
+     */
+    public Map<V, V> getForwardMapping(){
+        return Collections.unmodifiableMap(forwardMapping);
+    }
+
+    /**
+     * Get an unmodifiable version of the backward mapping function.
+     *
+     * @return the unmodifiable backward mapping function
+     */
+    public Map<V, V> getBackwardMapping(){
+        return Collections.unmodifiableMap(backwardMapping);
+    }
+
+    /**
+     * Get the active domain of the isomorphism.
+     *
+     * @return the set of vertices $v$ for which the mapping is defined
+     */
+    public Set<V> getMappingDomain(){
+        return forwardMapping.keySet();
+    }
+
+    /**
+     * Get the range of the isomorphism.
+     *
+     * @return the set of vertices $v$ for which a preimage exists
+     */
+    public Set<V> getMappingRange(){
+        return backwardMapping.keySet();
     }
 
     /**
@@ -127,10 +174,25 @@ public class IsomorphicGraphMapping<V, E>
     }
 
     @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        IsomorphicGraphMapping<?, ?> that = (IsomorphicGraphMapping<?, ?>) o;
+        return Objects.equals(forwardMapping, that.forwardMapping) &&
+                Objects.equals(backwardMapping, that.backwardMapping);
+    }
+
+    @Override
+    public int hashCode() {
+
+        return Objects.hash(forwardMapping, backwardMapping);
+    }
+
+    @Override
     public String toString()
     {
         StringBuilder str = new StringBuilder("[");
-        Set<V> vertexSet = g1.getGraph().vertexSet();
+        Set<V> vertexSet = graph1.vertexSet();
         Map<String, V> vertexMap = new TreeMap<>();
 
         for (V v : vertexSet) {
@@ -156,13 +218,59 @@ public class IsomorphicGraphMapping<V, E>
      */
     public boolean isEqualMapping(GraphMapping<V, E> rel)
     {
-        for (V v : g2.getGraph().vertexSet()) {
+        for (V v : graph2.vertexSet()) {
             if (!getVertexCorrespondence(v, false).equals(rel.getVertexCorrespondence(v, false))) {
                 return false;
             }
         }
 
         return true;
+    }
+
+    /**
+     * Computes the composition of two isomorphisms.
+     * Let $f : V_{G_1} \rightarrow V_{G_2}$ be an isomorphism from $V_{G_1}$ to $V_{G_2}$ and
+     * $g : V_{G_2} \rightarrow V_{G_3}$ one from $V_{G_2}$ to $V_{G_3}$.
+     *
+     * This method computes an isomorphism $h : V_{G_1} \rightarrow V_{G_3}$ from $V_{G_1}$ to $V_{G_3}$.
+     *
+     * Note: The composition $g ∘ f$ can be built only if $f$'s codomain equals $g$'s domain;
+     * this implementation only requires that the former is a subset of the latter.
+     *
+     * @param otherMapping the other isomorphism (i.e. function $g$)
+     * @return the composition of the two isomorphism
+     */
+    public IsomorphicGraphMapping<V, E> compose(IsomorphicGraphMapping<V, E> otherMapping){
+        Map<V, V> fMap = new HashMap<>(forwardMapping.size());
+        Map<V, V> bMap = new HashMap<>(forwardMapping.size());
+
+        for (V v: graph1.vertexSet()){
+            V u = otherMapping.getVertexCorrespondence(forwardMapping.get(v), true);
+            fMap.put(v, u);
+            bMap.put(u, v);
+        }
+
+        return new IsomorphicGraphMapping<>(fMap, bMap, graph1, otherMapping.graph2);
+    }
+
+    /**
+     * Computes an automorphism (i.e. an isomorphism mapping from a graph to itself).
+     *
+     * @param graph the input graph
+     * @param <V> the graph vertex type
+     * @param <E> the graph edge type
+     * @return a mapping from graph to graph
+     */
+    public static <V, E> IsomorphicGraphMapping<V, E> identity(Graph<V, E> graph){
+        Map<V, V> fMap = new HashMap<>(graph.vertexSet().size());
+        Map<V, V> bMap = new HashMap<>(graph.vertexSet().size());
+
+        for (V v: graph.vertexSet()){
+            fMap.put(v, v);
+            bMap.put(v, v);
+        }
+
+        return new IsomorphicGraphMapping<>(fMap, bMap, graph, graph);
     }
 }
